# algos_3assignment # MyLinkedListQueue or MyArrayListQueue ## *This class is an implementation of a queue using a LinkedList or ArrayList data structure. It has the following methods:*
enqueue(element): Adds an element to the back of the queue.

dequeue(): Removes and returns the front element of the queue.

peek(): Returns the front element of the queue without removing it.

size(): Returns the number of elements in the queue.

isEmpty(): Returns true if the queue is empty, false otherwise.

# MyLinkedListStack or MyArrayListStack ##
*This class is an implementation of a stack using a LinkedList or ArrayList data structure. It has the following methods:*

push(element): Adds an element to the top of the stack.

pop(): Removes and returns the element at the top of the stack.

peek(): Returns the element at the top of the stack without removing it.

size(): Returns the number of elements in the stack.

isEmpty(): Returns true if the stack is empty, false otherwise.